+++ src/events.test.ts
@@ -1,12 +1,10 @@
 import Axios from 'axios';
-import * as moment from 'moment';
 import {
   eventRequestDeadline,
   INTERVAL,
   pollIteration,
   requestEvents,
-  resetEventsPolling,
-  shouldRequestNotifications
+  resetEventsPolling
 } from 'src/events';
 
 Date.now = jest.fn(() => 1234567890);
@@ -42,35 +40,6 @@ describe('events module', () => {
           })
       );
       requestEvents();
-    });
-  });
-
-  describe('shouldRequestNotifications', () => {
-    const d1 = '2019-05-23T12:00:00';
-    const d2 = '2019-05-23T12:00:01';
-
-    const d1InMilliseconds = moment.utc(d1).valueOf();
-    const d2InMilliseconds = moment.utc(d2).valueOf();
-
-    it('should return `true` if there is a linode_resize event created AFTER the last time notifications were updated', () => {
-      expect(
-        shouldRequestNotifications(d1InMilliseconds, 'linode_resize', d2)
-      ).toBe(true);
-    });
-
-    it('should return `false` if there is a linode_resize event created BEFORE the last time notifications were updated', () => {
-      expect(
-        shouldRequestNotifications(d2InMilliseconds, 'linode_resize', d1)
-      ).toBe(false);
-    });
-
-    it("should return `false` for events that aren't related to notifications", () => {
-      expect(
-        shouldRequestNotifications(d1InMilliseconds, 'linode_boot', d2)
-      ).toBe(false);
-      expect(
-        shouldRequestNotifications(d1InMilliseconds, 'linode_update', d2)
-      ).toBe(false);
     });
   });
 });
