+++ src/store/linodes/linodes.events.ts
@@ -1,3 +1,4 @@
+import * as moment from 'moment';
 import { Dispatch } from 'redux';
 import { ApplicationState } from 'src/store';
 import { requestLinodeForStore } from 'src/store/linodes/linode.requests';
@@ -8,6 +9,14 @@ import { deleteLinode, updateLinode } from './linodes.actions';
 const linodeEventsHandler: EventHandler = (event, dispatch, getState) => {
   const { action, entity, status } = event;
   const { id } = entity;
+
+  // We may want to request notifications here, depending on the event
+  // action, time of creation, and the last time we updated notifications.
+  const { lastUpdated } = getState().__resources.notifications;
+
+  if (shouldRequestNotifications(lastUpdated, event.action, event.created)) {
+    dispatch(requestNotifications() as any);
+  }
 
   switch (action) {
     /** Update Linode */
@@ -151,3 +160,40 @@ const handleLinodeCreation = (
       return;
   }
 };
+
+/**
+ * shouldRequestNotifications
+ *
+ * We may want to request notifications upon receiving an event, depending on the
+ * type of the event, and the last time we requested notifications.
+ *
+ * EXAMPLE:
+ *
+ * 1) User clicks "Resize Linode"
+ * 2) There is now a `migration_imminent` notification on the Linode.
+ * 3) In the next few minutes, the resize is kicked off.
+ * 4) The `migration_imminent` notification on the Linode goes away.
+ */
+export const shouldRequestNotifications = (
+  notificationsLastUpdated: number,
+  lastEventAction?: Linode.EventAction,
+  lastEventCreated?: string
+) => {
+  if (!lastEventAction || !lastEventCreated) {
+    return false;
+  }
+
+  return (
+    eventsWithRelevantNotifications.includes(lastEventAction) &&
+    // if the event was created after the last time notifications were updated
+    moment.utc(lastEventCreated).isAfter(moment.utc(notificationsLastUpdated))
+  );
+};
+
+const eventsWithRelevantNotifications: Linode.EventAction[] = [
+  'linode_resize',
+  'linode_resize_create',
+  'linode_migrate',
+  'linode_mutate',
+  'linode_mutate_create'
+];
